Debug ADC changed setup  SPI UART

+++ Inc/spi_adc.h
@@ -23,7 +23,7 @@
 
 int32_t SPI_Get_RAW_data_ADC7767 ( SPI_TypeDef *SPIx );
 
-int16_t convert_RAW_data_ADC_24b_to_16b( int32_t raw_data_adc_24b, uint8_t Vref_adc, _SETTINGS_MODULE *config_module );
+int16_t convert_RAW_data_ADC_24b_to_16b( int32_t raw_data_adc_24b, float  Vref_adc, _SETTINGS_MODULE *config_module );
